// src/app/shared/enums/drawing-modes.ts
export enum DrawingModes{
    BRUSH_MODE = 'brush',
    INTERACTIVE = 'interactive',
    TEXT_BRUSH = 'text brush',
    CIRCLE_MODE = 'circle',
    HOOK_BRUSH = 'force brush',
    CLEAR = 'ClearBrush',
    GRADIENT_LINE = 'gradient line',
    BEADS = 'beads'
}

export enum OtherModes{
    NOISE = 'noise',
    GAN = 'GAN'
}